+++ Gems/Atom/Tools/AtomToolsFramework/Code/Include/AtomToolsFramework/Window/AtomToolsMainWindow.h
@@ -8,6 +8,7 @@
 
 #pragma once
 
+#include <AtomToolsFramework/AssetBrowser/AtomToolsAssetBrowser.h>
 #include <AtomToolsFramework/Window/AtomToolsMainWindowRequestBus.h>
 #include <AzCore/Memory/SystemAllocator.h>
 #include <AzQtComponents/Components/DockMainWindow.h>
@@ -52,6 +53,8 @@ namespace AtomToolsFramework
         QMenu* m_menuEdit = {};
         QMenu* m_menuView = {};
         QMenu* m_menuHelp = {};
+
+        AtomToolsFramework::AtomToolsAssetBrowser* m_assetBrowser = {};
 
         AZStd::unordered_map<AZStd::string, AzQtComponents::StyledDockWidget*> m_dockWidgets;
         AZStd::unordered_map<AZStd::string, QAction*> m_dockActions;
