+++ Gems/Atom/Tools/AtomToolsFramework/Code/Source/AssetBrowser/AtomToolsAssetBrowser.ui
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <ui version="4.0">
- <class>MaterialBrowserWidget</class>
- <widget class="QWidget" name="MaterialBrowserWidget">
+ <class>AtomToolsAssetBrowser</class>
+ <widget class="QWidget" name="AtomToolsAssetBrowser">
   <property name="geometry">
    <rect>
     <x>0</x>
