+++ Gems/Atom/Tools/MaterialEditor/Code/Source/Window/MaterialEditorWindowModule.cpp
@@ -14,6 +14,7 @@ void InitMaterialEditorResources()
     //Must register qt resources from other modules
     Q_INIT_RESOURCE(MaterialEditor);
     Q_INIT_RESOURCE(InspectorWidget);
+    Q_INIT_RESOURCE(AtomToolsAssetBrowser);
 }
 
 namespace MaterialEditor
