+++ Gems/Atom/Tools/ShaderManagementConsole/Code/Source/Window/ShaderManagementConsoleWindow.h
@@ -14,9 +14,7 @@
 #include <AtomToolsFramework/Document/AtomToolsDocumentMainWindow.h>
 
 AZ_PUSH_DISABLE_WARNING(4251 4800, "-Wunknown-warning-option") // disable warnings spawned by QT
-#include <Window/ShaderManagementConsoleBrowserWidget.h>
 #include <Window/ToolBar/ShaderManagementConsoleToolBar.h>
-
 #include <QStandardItemModel>
 AZ_POP_DISABLE_WARNING
 #endif
@@ -40,6 +38,6 @@ namespace ShaderManagementConsole
     protected:
         QWidget* CreateDocumentTabView(const AZ::Uuid& documentId) override;
 
-        ShaderManagementConsoleToolBar* m_toolBar = nullptr;
+        ShaderManagementConsoleToolBar* m_toolBar = {};
     };
 } // namespace ShaderManagementConsole
